Refactor and fix documentations.
