+++ README.md
@@ -28,9 +28,9 @@ All APIs are TLS-protected, hence please use `https://` or `wss://`.
 Defined in the [OpenAPI specification](./openapi.yaml).
 
 ### WebSocket APIs
-#### /ws-notarize
+#### /notarize
 ##### Description
-To perform notarization using the session id (unique id returned upon calling the `/notarize` endpoint successfully) submitted as a custom header.
+To perform notarization using the session id (unique id returned upon calling the `/session` endpoint successfully) submitted as a custom header.
 
 ##### Custom Header
 `X-Session-Id`
@@ -54,7 +54,7 @@ To perform notarization, some parameters need to be configured by the prover and
 - maximum transcript size
 - unique session id
 
-To streamline this process, a single HTTP endpoint (`/notarize`) is used by both TCP and WebSocket clients. The only difference being, for TCP client, the notarization process will be kickstarted at the end of this configuration, whereas WebSocket client will need to establish a separate WebSocket connection to a different endpoint (`/ws-notarize`).
+To streamline this process, a single HTTP endpoint (`/session`) is used by both TCP and WebSocket clients.
 
 #### WebSocket
-Axum's internal implementation of WebSocket uses [tokio_tungstenite](https://docs.rs/tokio-tungstenite/latest/tokio_tungstenite/), which provides a WebSocket struct that doesn't implement [AsyncRead](https://docs.rs/futures/latest/futures/io/trait.AsyncRead.html) and [AsyncWrite](https://docs.rs/futures/latest/futures/io/trait.AsyncWrite.html). Both these traits are required by TLSN core libraries for prover and notary. To overcome this, a [slight modification](./src/axum_websocket.rs) of Axum's implementation of WebSocket is used, where [async_tungstenite](https://docs.rs/async-tungstenite/latest/async_tungstenite/) is used instead so that [ws_stream_tungstenite](https://docs.rs/ws_stream_tungstenite/latest/ws_stream_tungstenite/index.html) can be used to wrap on top of the WebSocket struct to get AsyncRead and AsyncWrite implemented.
+Axum's internal implementation of WebSocket uses [tokio_tungstenite](https://docs.rs/tokio-tungstenite/latest/tokio_tungstenite/), which provides a WebSocket struct that doesn't implement [AsyncRead](https://docs.rs/futures/latest/futures/io/trait.AsyncRead.html) and [AsyncWrite](https://docs.rs/futures/latest/futures/io/trait.AsyncWrite.html). Both these traits are required by TLSN core libraries for prover and notary. To overcome this, a [slight modification](./src/service/axum_websocket.rs) of Axum's implementation of WebSocket is used, where [async_tungstenite](https://docs.rs/async-tungstenite/latest/async_tungstenite/) is used instead so that [ws_stream_tungstenite](https://docs.rs/ws_stream_tungstenite/latest/ws_stream_tungstenite/index.html) can be used to wrap on top of the WebSocket struct to get AsyncRead and AsyncWrite implemented.
