+++ src/domain/notary.rs
@@ -6,18 +6,18 @@ use tokio::sync::Mutex;
 
 use crate::config::NotarizationProperties;
 
-/// Response object of the /notarize API
+/// Response object of the /session API
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct NotarizationResponse {
+pub struct NotarizationSessionResponse {
     /// Unique session id that is generated by notary and shared to prover
     pub session_id: String,
 }
 
-/// Request object of the /notarize API
+/// Request object of the /session API
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct NotarizationRequest {
+pub struct NotarizationSessionRequest {
     pub client_type: ClientType,
     /// Maximum transcript size in bytes
     pub max_transcript_size: Option<usize>,
@@ -45,12 +45,11 @@ impl NotaryGlobals {
     pub fn new(
         notary_signing_key: SigningKey,
         notarization_config: NotarizationProperties,
-        store: Arc<Mutex<HashMap<String, Option<usize>>>>,
     ) -> Self {
         Self {
             notary_signing_key,
             notarization_config,
-            store,
+            store: Default::default(),
         }
     }
 }
